libs/adk-agents/pyproject.toml
[project]
name = "adk-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Julian Hecker" }]
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.11.0",
    "google-genai>=1.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
